Rename browser folder to browserContext
