+++ packages/interbit-ui-tools/src/middleware/browserContext/staticHtml.js
